requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
faiss-cpu==1.7.4
filelock==3.13.1
Flask==3.0.0
flask-restx==1.3.0
flask-swagger-ui==4.11.1
fsspec==2023.12.2
greenlet==3.0.3
gunicorn==21.2.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psutil==5.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.16.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.24
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
